services/sensor-data-service/requirements.txt
paho-mqtt==1.6.1
influxdb-client==1.41.0
requests==2.32.3
